Add spring security

+++ src/main/java/com/controller/board/BoardController.java
@@ -1,5 +1,6 @@
 package com.controller.board;
 
+import java.security.Principal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -7,7 +8,6 @@ import javax.annotation.Resource;
 import javax.servlet.ServletRequest;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,9 +28,9 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.biz.board.BoardService;
 import com.biz.board.BoardVO;
 import com.common.paging.PageMaker;
-import com.common.validation.UpdateBoardValidator;
 import com.common.vo.FileVO;
 import com.common.vo.SearchVO;
+import com.exception.validation.UpdateBoardValidator;
 
 @Controller
 public class BoardController {
@@ -49,19 +49,13 @@ public class BoardController {
 	
 	@RequestMapping(value="/insertBoard.do", method=RequestMethod.POST)
 	@ResponseBody
-	public ResponseEntity<Map<String,Object>> insertBoard(HttpSession session, @Valid @ModelAttribute BoardVO vo, 
+	public ResponseEntity<Map<String,Object>> insertBoard(@Valid @ModelAttribute BoardVO vo, 
 			BindingResult brs) throws MethodArgumentNotValidException{
-		Map<String, Object> map = new HashMap<String, Object>();
-		
-		// 로그아웃 상태에서 글 등록 불가.
-		if(session.getAttribute("user") == null){
-			map.put("location", "login.do");
-			return new ResponseEntity<Map<String, Object>>(map, HttpStatus.OK);
-		}
-		
 		if(brs.hasErrors()){ throw new MethodArgumentNotValidException(null, brs); }
 		
-		Integer boardId = boardService.insertBoard(vo, basePath);
+		Map<String, Object> map = new HashMap<String, Object>();
+		
+		int boardId = boardService.insertBoard(vo, basePath);
 		map.put("location", "getBoard.do?boardId="+boardId);
 		return new ResponseEntity<Map<String,Object>>(map, HttpStatus.OK);
 	}
@@ -79,11 +73,12 @@ public class BoardController {
 	// 글 목록 조회
 	@RequestMapping(value="/getBoardList.do")
 	public String getBoardList(SearchVO search, @RequestParam(defaultValue="1") int curPage, Model model, ServletRequest request) {
-		/* WebApplicationContext TEST
+		/*// WebApplicationContext TEST
 		WebApplicationContext ctx = ContextLoader.getCurrentWebApplicationContext();	// ContextLoaderListener 이용해서 생성한 WebApplicationContext
 		WebApplicationContext ctx2 = RequestContextUtils.getWebApplicationContext(request);	// DispatcherServlet 이용해서 생성한 WebApplicationContext
-  		ctx.hashCode() == ctx2.getParent().hashCode() == true
-		*/
+  		// ctx.hashCode() == ctx2.getParent().hashCode() == true
+		String[] strs = ctx2.getBeanDefinitionNames();
+		for(String str : strs) System.out.println(str);*/
 		
 		if(search.getSearchCondition() == null) search.setSearchCondition("TITLE");
 		if(search.getSearchKeyword() == null) search.setSearchKeyword("");
@@ -108,7 +103,12 @@ public class BoardController {
 	
 	// 글 수정
 	@RequestMapping(value="/updateBoard.do", method=RequestMethod.GET)
-	public String updateBoard(BoardVO vo, Model model){
+	public String updateBoard(BoardVO vo, Model model, Principal principal){
+		// 게시글 작성 아이디와 로그인 아이디가 다른 경우
+		// @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ security 처리 가능하면 security로 처리 or 예외처리.
+		if(!principal.getName().equals(vo.getUserId())){ 
+			return "redirect:getBoard.do?boardId="+vo.getBoardId(); 
+		}
 		model.addAttribute("boardMap", boardService.getUpdateBoard(vo));
 		return "board/updateBoard.jsp";
 	}
@@ -129,7 +129,12 @@ public class BoardController {
 	
 	// 글 삭제
 	@RequestMapping(value="/deleteBoard.do")
-	public String deleteBoard(BoardVO vo){
+	public String deleteBoard(BoardVO vo, Principal principal){
+		// 게시글 작성 아이디와 로그인 아이디가 다른 경우
+		// @@@@@@@@@@@@@@@@@@@@@@@@@@@@@ security 처리 가능하면 security로 처리 or 예외처리.
+		if(!principal.getName().equals(vo.getUserId())){ 
+			return "redirect:getBoard.do?boardId="+vo.getBoardId(); 
+		}
 		boardService.deleteBoard(vo);
 		return "redirect:getBoardList.do";
 	}
@@ -140,6 +145,4 @@ public class BoardController {
 	public  ResponseEntity<byte[]>  fileDownload(HttpServletRequest request,HttpServletResponse response,FileVO vo) {
 	   	return boardService.downloadFile(request, response, vo);
 	}
-	
-	
 }
